[HOTSPOT-216] 가족 구성원 삭제 신청 로직 수정 & 요금제 데이터 사용량 데이터 타입 변경

+++ src/main/java/hotspot/user/plan/domain/Plan.java
@@ -13,6 +13,6 @@ import lombok.Getter;
 public class Plan {
     private final Long id; // DB 저장 전에는 null, 저장 후 ID 할당됨
     private final String name;
-    private final int dataAmount;
+    private final Long dataAmount;
     private final DataPeriod dataPeriod;
 }
